pyproject.toml
[tool.poetry]
name = "nubi-challenge"
version = "0.1.0"
description = ""
authors = ["EmannuelDM <ema.dmaccio@gmail.com>"]
readme = "README.md"
packages = [{include = "nubi_challenge"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.103.2"
sqlalchemy = "^2.0.22"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
